// generated/openweather-schemas/forecastResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * OpenWeatherMap API
 * Weather data API
 * OpenAPI spec version: 2.5.0
 */
import type { ForecastItem } from "./forecastItem";
import type { City } from "./city";

export interface ForecastResponse {
  cod?: string;
  message?: number;
  cnt?: number;
  list?: ForecastItem[];
  city?: City;
}