// generated/weatherapi-schemas/realtimeWeather200One.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Weather API
 * # Introduction
WeatherAPI.com provides access to weather and geo data via a JSON/XML restful API. It allows developers to create desktop, web and mobile applications using this data very easy. We provide following data through our API:    
- Real-time weather
- 14 day weather forecast
- Historical Weather
- Marine Weather and Tide Data
- Future Weather (Upto 365 days ahead)
- Daily and hourly intervals
- 15 min interval (Enterprise only)
- Astronomy
- Time zone
- Location data
- Sports
- Search or Autocomplete API
- Weather Alerts
- Air Quality Data
- Bulk Request

# Getting Started  

You need to [signup](https://www.weatherapi.com/signup.aspx) and then you can find your API key under [your account](https://www.weatherapi.com/login.aspx), and start using API right away!

Try our weather API by using interactive [API Explorer](https://www.weatherapi.com/api-explorer.aspx).

We also have SDK for popular framework/languages available on [Github](https://github.com/weatherapicom/) for quick integrations.

If you find any features missing or have any suggestions, please [contact us](https://www.weatherapi.com/contact.aspx).  

# Authentication  

API access to the data is protected by an API key. If at anytime, you find the API key has become vulnerable, please regenerate the key using Regenerate button next to the API key.  

Authentication to the WeatherAPI.com API is provided by passing your API key as request parameter through an API .    

##  key parameter 
key=YOUR API KEY 

 * OpenAPI spec version: 1.0.2
 */
import type { Location } from "./location";
import type { Current } from "./current";

export type RealtimeWeather200One = {
  location?: Location;
  current?: Current;
};